### 2024-12-17 #7600 @editorialbot
[PRE REVIEW]: OpenTTDLab: A Python framework for reproducible experiments using OpenTTD
**Submitting author:** @michalc (<a href="http://orcid.org/0009-0009-1002-1036">Michal Charemza</a>)
**Repository:** https://github.com/michalc/OpenTTDLab
**Branch with paper.md** (empty if default branch): main
**Version:** v0.0.74
