### 2025-01-04 f3ba5c9
feat: support OpenTTD 15
This fixes https://github.com/michalc/OpenTTDLab/issues/234 by using the
"schedule" console script command, added in
https://github.com/OpenTTD/OpenTTD/pull/12761 and released in OpenTTD 15, to
emulate the pre OpenTTD 14 in-game-time autosave behaviour.

This also adds in a savegame just after the AI creation, which I think is
probably a good thing in terms of charting results
