### 2025-01-05 00b52e4
docs: add missing comma to BibTex entry
