### 2025-01-05 8a122e8
docs: don’t abbreviate United Kingdom in paper
