+++ paper/paper.bib
@@ -27,7 +27,7 @@
     month        = {11},
     year         = {2024},
     school       = {University of Edinburgh},
-    address      = {UK},
+    address      = {United Kingdom},
     url          = {https://doi.org/10.7488/era/5414},
     doi          = {10.7488/era/5414},
 }
