### 2025-03-27 #7600 @skanwal
Hello @SunnyXu and @rwood-97 - are you available to review this submission for JOSS?
