### 2025-04-28 #8014 @harkeganesh
@skanwal
I’ve completed my review of *OpenTTDLab* and recommend acceptance with one recommendation on the docstring. The software is well-designed for reproducible *OpenTTD* experiments, with clear documentation (in [README.md](http://readme.md/)), examples, and tests for core functionality. The software runs and installs as expected. Example notebooks work as expected. The paper is well-written, meets JOSS scope, and appropriately addresses the statement of need. The repository includes an OSI-approved License, and the single author has made substantial contributions.

**Strengths**:

- Clear statement of need and target audience (researchers, AI developers).
- Comprehensive documentation.
- The paper is clear, well-structured, and adequately references the field.

**Recommendations**:

1. **Python Code Documentation**: Improve Python code documentation with docstrings.

Please let me know if you have any questions. Thank you.
