### 2025-04-29 #8014 @CalebAbhulimhen
Overall, I'm impressed with the quality of this project, including the paper, extensive documentation, and novelty of the idea. I believe this software paper fits all the criteria needed, so I'm also **+1 on its acceptance**.

I have a proposal - echoing @harkeganesh - it will be nice to add some documentation to the Python code to make it a bit more readable, and also restructure the codebase to follow the [conventional Python packaging recommendation](https://packaging.python.org/en/latest/tutorials/packaging-projects/#creating-the-package-files). While this isn't strictly a JOSS criteria, making these changes will make it easier for other developers to easily navigate and contribute to this amazing proiect.

This is my review for this paper, handing off to @skanwal. Thanks!
