### 2025-04-29 #8014 @michalc
> Double check authors and affiliations (including ORCIDs)

Done - just me. Although I'm not 100% if my affiliation is appropriate: I did the bulk of the work while I was a student at the University of Edinburgh for a dissertation, but I am no longer a student or otherwise affiliated with it.

> Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.

No changes from the review (I see there are recommendations, but it sounds like they are not blockers), so the version is v0.0.75 https://github.com/michalc/OpenTTDLab/releases/tag/v0.0.75

> Archive the release on Zenodo/figshare/etc and post the DOI here.

https://doi.org/10.5281/zenodo.15305719

> Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.

Done

> Make sure that the license listed for the archive is the same as the software license.

Done - GNU General Public License v2.0
