### 2025-06-04 #8014 @michalc
Thank you @crvernon, @skanwal, @CalebAbhulimhen and @harkeganesh so much for all the time you put into this!!

And thank you to everyone behind the scenes as well... both for any work relating to my paper directly if there was any, but also for all the work before I even submitted: I imagine a _lot_ of effort has been put into JOSS over the years (e.g. editorial bot, the various sites, etc etc)

Thank you!!
